@@ -2,7 +2,7 @@ apply plugin: 'com.android.library'
 
 android {
 
-    compileSdkVersion 33
+    compileSdkVersion 35
     namespace "com.theartofdev.edmodo.cropper"
     defaultConfig {
         minSdkVersion 14
